requirements.txt
numpy
openai
pip
python-dotenv
PyYAML
graphviz
networkx
pyvis
pandas==1.5.3